refactor: Extract quick jumping logic to ContainerQuickJumpModel
